Fixes loop index type
Sets it to size_t vs int, since that's what we're comparing.

Closes #2663

+++ steamshovel/private/steamshovel/ScenarioFrameModel.cpp
@@ -361,7 +361,7 @@ void ScenarioFrameModel::inputsChanged( I3FramePtr ptr ){
 		row_data_.push_back( dec );
 	}
 
-	for( int i = 0; i < frame_keys.size(); ++i ){
+	for( size_t i = 0; i < frame_keys.size(); ++i ){
 		RowDatum* this_row = NULL;
 		// a keylist of just this key
 		ArtistKeyList keys( 1, frame_keys.at(i) );
